Peer operations suspend call is in right place for sharding
